Added E-field saving when function name in save_list as well as a check on the planes that were actually saved
Previously you'd have to add 'if plane_name is not None: wfo.save_plane(plane_name)' to the end of every function in a prescription you think the user would want to save the plane for.

The check just ensures that sp.save_list and self.save_planes are the same. if not the user could've put a typo in save_list and so that locaiton was missed