cpx_sequence is now intelligently saved or as the user specifies. Detailed explanation below...
Initially cpx_sequence was saved in one go at the end of the simulation. This method lacks checkpointing and runs into memory issues for wide/long simulations.

sp.memory_limit is number of bytes for sixcube of complex fields before chunking happens and sp.checkpointing (int or None) is the number of timesteps before complex fields sixcube is saved. The minimum steps associated with these two is used for chunking in the time domain

These chunks can be saved to npy in realtime or used to populate a larger array in memory. Default is no checkpointing/chunking and just load all into memory

Also added sp.verbose and sp.debug options to turn off printing and plotting as neccessary