The RunMedis class now in the initialisation takes care of checking the parameters in the config match and if not creates the new directories.
The simulated data product, either telescope or the detector is now a property of the class

Similarly Telescope and (not implemented yet but later MKID) are there own classes with initialisations (eg do the atmospheres exist, no? okay create them) and a call() (create the sequence of electric fields I prefer the name "fields" over "cpx_sequence")

The cache data for each test simulaition is now a pickle of the params used, atmos maps, aberration maps and a copy of the prescription (as a backup in case the prescription gets updated and the simulation performed using it won't match one performed on the original under the params.py configuration)

I vastly simplifed the multiprocessing

Save and load of the electric fields are incorporated into this class too

lens_params is now a list of dictionaries (one dictionary for each lens). Each dictionary contrains eg the name, focal length, diam and aberation_vals. When initialising the telescope it looks through this list

I pretty strongly believe we should move away from global parameter values imported at the top of each module that can be updated. On so many occasions i've run into errors because some parameter wasn't updated when it should have been in a namespace or vice versa (not to mention all the times this might be happening and I don't realise). Passing only the neccessary variables should make functionss more atomic also